Added support for variables with UNDEFINED vertical levels. They are now treated as a 'one-level' variable - meaning they are all glommed together. The matlab scripts also support UNDEFINED.
And, Just For Nancy, I fixed the logic of the program so that now
if you have JUST ONE observation sequence file in another directory,
obs_diag will process it and terminate normally rather than die if
the directory node does not have an 'obvious' way to increment it.
i.e.:
   obs_sequence_name = '/Users/thoar/temp/obs_seq.final',
works.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@3457 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: bf2d4b9b06d45ac00001f5873c7dfbd5fbabfe07