Make public the routine which changes a generic kind string into an index number (the corresponding routine for specific types has existed for a long time; but there was nothing for the kinds). This is needed by josh for the wrf model_mod revisions.  The original routines were contributed by tim whitcomb who also needed them for the COAMPS model.
Also, a group of changes which are internal to the code and
*not* externally visible: renaming, where possible, 'kind' into
'generic' and 'type' into 'specific', to start to try to ease
the eventual transition from 'kind' in the public routine names
and arguments into 'type'.  Right now most of the public usage
is incorrect, which contributes to the confusion about which
is which.  This requires a one line change to the preprocess
program to rename an automatically built table.


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@3764 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 6a9c458789b935ae7a5dc3561d0f8a6e642b0e59