Update to make smoother work correctly.  All files in this commit batch will need to be updated together in order to build, as the move_ahead() function has been split into 2 parts to accomodate the need for updating at the proper time when smoothing.
Also added better tracing facility to filter - set trace_execution to true
in the &filter_nml in input.nml to get detailed information about what part
of the cycle is being executed.  You will also get more details about what 
the data time is and what the model advance time should be.

Updated namelists and html to match current code.

Smoother fixed to use proper states in circular buffer, to not overwrite
times with incorrect values, to handle runs where the first action is
assimilation and not a model advance, and to not advance when the main
execution loop is being exited because all obs have been assimilated.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@3799 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 5ac1965cf972d1e98dc5a4371155d86f86050aa3