Everything compiles on coral with Intel 10.1.
The pop_to_dart code runs and is to be thoroughly checked by the
matlab script Check_pop_to_dart.m. At present, I think matlab
unwraps the N-D variables in opposite order to the way they are
handled in the Fortran code.

The model_mod:static_init_model() routine checks out, but the
stubs for orthogonality are just that - stubs.

The communication of how often it is possible to stop the model
and when to generate the restart file is not finished.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@3945 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 50063a96ec4d8f79f9e45479eac0804e1fd4849a