Updated wrf model_mod:
Contains Ryan's new vortex center-finding code and a max 
surface height difference limit.  The default is to use the old code
so it's backwards compatible; edit the code and set 'use_old_vortex'
to .false. to try the new vortex code.  The default for the surface
height limit is -1, which disables the check so it's also backwards
compatible.

New namelist items:  sfc_elev_max_diff, circulation_pres_level, and
circulation_radius.  see docs for usage.

The routine which tests the height difference between the model surface
and the elevation of a surface observation is called 'height_diff_check'
and is public.

The sizes of local variables inside the spline routines have been
increased.

The old vortex code has changed in one place - if the center is located
on any of the edges of the search box, the forward operator will return
a failure code instead of returning success and the same location as
the incoming observation.

Other changes:

input.nml - the radar section has been updated; _r8 is not legal on
numeric constants in a namelist file.  each namelist is terminated
by a slash on a separate line in a consistent column.  two new
sections have been added:  &dart_to_wrf_nml and &replace_wrf_fields_nml.
see the docs for use.

dart_to_wrf - requires a namelist now.  the namelist settings allow
restart files to be converted to wrf netcdf files for continuing with
a free run after an assimilation.

replace_wrf_fields - a fortran version of the ncl scripts which
take 2 netcdf files and replace the fields from one into the other.
namelist options control which fields are copied, and whether a
missing field is simply skipped or becomes a fatal error.  the
field names can either be specified directly in the namelist or
can be stored in a text file, one fieldname per line.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@4283 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: cc00535ac86680774df6e6accab670b99d671bd8