The 'time_to_skip' is correctly described as being the amount of time FROM THE LEADING EDGE OF THE FIRST BIN - not the bin 'center'. The logic of the programs has not changed, just the documentation. The matlab plotting scripts also act on the time appropriately.
Additionally, the annotation of the input file at the bottom of the 
plot is now more robust and will correctly print the full name of
the file no matter if the file is in the MATLABPATH, the local working
directory, or is an absolute path.

The obs_diag.f90 summary statistics (written to stdout & logfile) have
a bit more verbosity about what sorts of DART QC flags were encountered.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@4352 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: df8ed4de06a23ec888da2ff523eb49e1a34df05b