The scripts now appropriately put the land mask at the top or bottom, depending on the the 'which_vert' portion of the location AND some knowledge of the origin of the observation. At present observations of oceanic quantities (i.e. observations with a 'WOD observation' copy) get plotted with a Z label of 'depth' even though their which_vert is 'height' ... pretty tricksy. Atmospheric quantities are handled correctly, as far as I have been able to test.
The figure with multiple panels is simpler, and the bottom graphic
is generally a plot of the observation value vs. the ensemble mean estimate.
If these things are alike, they line up on the diagonal.

The summary of how many obs got rejected, etc. is ONLY echoed to 
the command window.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@4356 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 79c37d0b65988f0cbe3bec6eee4c71ad630ee794