Adding a link to David Dowell's radar preprocessing code. (Here and in the DART:Miscellany:Useful Software tab)
On Feb 22 Nancy wrote:
"last week in the wrf/dart meeting david d gave a
presentation on his radar preprocessing code.
he's uploaded the software and constructed doc
pages and whatnot.  the web link is:

http://code.google.com/p/opaws/

he's got a pointer to dart on that page, and we
said we'd add a pointer to his page in our stuff.

i'll add one on our developer's wiki page in the
wrf section, but if there's a place on our image
web site where it might fit that would be good
to add, too."



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@4852 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 842a5a305fe03a453651194eacb58cd416af4ef4