"model_interpolate" now uses the gravity-adjusted geopotential height computed by TIEGCM for vertical interpolation. In order to pass the geopotential height specific to each ensemble member to the "model_interpolate" routine, the DART state vector includes the gravity-adjusted 3D geoponetial height fields as auxiliary variables. Note that the updated geopotential height fields are not passed back to TIEGCM.
"get_state_meta_data" uses the ensemble mean of the gravity-adjusted geopotential height to assign height 
information to each element of the DART state vector.  TIEGCM's natural vertical coordinate is pressure.
For the sake of vertical localization that requires the computation of distance between obs and states, 
it is easier to use height as vertical coordinates for now.

The order in which 3D TIEGCM fields are placed in the DART state vector is changed to accommodate 
the usage of the module for assimilation of neutral density observations only (Use "only_neutral_density" switch).

Set "estimate_parameter" to be true to update model 1D parameters, when 1D paramters are included as part of
the DART state vector (i.e., state_num_1d > 0). Otherwise model 1D parameters will not be estimated.


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@5088 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 3a31562f63bae69978ef4fc5827bdd09e896b6d7