Same commit has been made on trunk.
Patch for the computation of correlation for inflation/sampling error
when running with R4 precision.  With very small values it was possible
for both A and B to test larger than 0, but A*B was small enough to 
underflow and be 0.  The original code was doing the tests separately
and not testing the product before dividing by it. With very small vals
it was possible to divide by 0 and get NaN or Inf (depending on the numerator).  
Inf would get reset to 1.0 in the lines below but NaN would remain as
NaN and percolate through the code to the output state vector unless you 
were running with floating point exceptions enabled (which usually carries 
a performance penalty).

This revised code checks the values in a slightly different order to
avoid computation if we are going to be setting the correlation to 0
anyway, and it is more careful to check for 0s in the face of underflow
with R4 precision.  There is still a risk of overflow in this code with
very large data values; we do not recommend running with reals compiled
R4 if your data includes values with very large (eg 1e15+) values.


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/releases/Kodiak@5577 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: d1cbd8518754f2dcad98de55c65e6c4055bd9c81