Apply updates which have been in the development version for a while. here is the original log msg:
Significant winds come in with vertical units of height in meters.
Use a variable with the name height instead of reusing the pressure
array.  Also, the conversion routine from height to pressure, in 
order to get the correct obs error, only works up to 44km and then
gives a segfault.  Limit the largest height going into the conversion
routine to 40km.  Also add a check to be sure the height isn't missing
before trying to use it.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/releases/Kodiak@5596 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 8e27962a2b41d4974c11fb0da327e106a0d0509a