fix error in moisture obs locations - they are not located on the levels. instead, they are actually the mean of the layers between the fixed levels. (levels, layers are the AIRS terminology for these things.)
temperature obs were ok before - located on the levels - but moisture obs
have changed location in this version of the converter - they are now located
at the midpoint, in log space, between the levels.

other housekeeping: changed mss commands in the scripts to use the hsi interface;
removed unused variables; made sure constants are _r8; updated documentation
to describe the location issues with the moisture obs.

git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@5923 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: fe21b735ff9ecf25bb4691b1198336a9d848c96d