rename the MODIS directory to 'tpw' since this converter starts with ASCII input files instead of reading the HDF-EOS MODIS files.  rework the code to have more namelist values. update the superobs code to wrap correctly around the prime meridian, and average the values in time as well (instead of using the time of the first obs being averaged).  write documentation and leave a pointer to a single data file which can be used for testing.
updated the docs for several of the other observation
converters. the snow doc was basically a copy of the
text_to_obs docs without any changes.  the quikscat
docs had the sense of the thinning backwards (the code
is fine; the description was misleading).  the AIRS
docs didn't include a namelist section at all.

also starting to convert the namelists into
the &bob a=1 b=2 format instead of namelist /bob/ a, b


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@6275 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: c302343345571dd86dd80362ff1e50c4fd01462c