Moved the 'cd to tempdir' to after the time determination which allows us to use the same relative filename for the pointer file to use to determine the time. (in both the CESM-flavor and individual model scripts)
Enabled the use of a cookie file for CAM startup to flag
the first model advance and possible creation of initial
inflation files so you don't have to have initial inflation
files before you startup - and can continue to have the
input.nml 'inf_initial_from_restart' TRUE for the entire 
experiment.

Set the configured ensemble size in the namelists that
are staged in the CASEROOT directory and removed it from
the assimilate.csh scripts. Less mysterious.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@6282 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 821c180fd63a919f5849c025b77f3f9be100549b