make the mkmf perl script generate a makefile target that will silently check the status of the mpi source files and alter them if they are configured for a different compiler.
fixsystem now takes a single argument: the name of the compiler
command, such as 'ifort', 'gfortran', 'pgf90', 'xlf90', etc.
it will ensure the source is compatible for that compiler.
for now for backwards compatibility this script can still be
called without any arguments; but this will eventually be
deprecated and the compiler name will have to be supplied
as an argument.

remove mention of fixsystem from the top level README.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@6377 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: e73d152c89b78505a6125ed0e29765d7709f21a8