have the rank histogram filter call a local norm_inv() routine instead of the stats toolbox norminv().
have the oned_model call a local kurt() routine
instead of the stats toolbox kurtosis().

have plot_gaussian call a local norm_pdf() routine
instead of the stats toolbox normpdf().

have the run_lorenz_96 code call different increment
routines based on the menu button for EAKF, EnKF, or RHF.

add a new norm_inv() routine which computes the
inverse CDF for a normal distribution, either for
a single scalar or for an array.  based on the
fortran code found in assim_tools_mod.f90.

with this update, DART_LAB should be completely
independent of the statistics toolbox.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@6831 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 1aa2a150820a1ae4b646eeb5451371ae69575a62