bug fix from trunk:
add an extra set of parens around an integer expression 
of the form:  a / b * b which has always been evaluated
(a / b) * b as expected, doing integer math roundoff.
however, one recent compiler (NAG) apparently evaluated 
it as: a / (b * b), possibly because b is a constant and 
more aggressive optimization could evaluate it early.
the result here was always 0 instead of the expected
roundoff expression.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/releases/Lanai@6905 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 07f751018d36e66603ee05ec6f6e9bda3c25d97e