Adding setup scripts for CESM1_2_1 There is a new interp_mod.f90 for cam/se with a bugfix. Made the 'sed' command more specific when updating the BSUB quantities. Made it easier to modify the RESTARTDIR in stage_cesm_files, if the 'rest' directory is actually a '.sta' directory, you can make that change in one place instead of three.
Added code to use the environment variable for MPI_RUN_COMMAND instead
of hardcoding one in CESM_DART_config ... made it more platform-agnostic.

Added the default value for the cam_phis namelist variable.
This value needs to change for different cores, so if it present in
the namelist, we have a hope of changing it automatically.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@6924 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 44ed2ac29a9714737c97822077a9d1d0cdc303a9