change the 'list' option to require N separate lists if you are going to compare more than a single set of obs_seq.final files.  each list contains the filenames for each experiment.  this is easier for the user to generate and less confusing all around.  fixes issue DART-30.
also support a new namelist item: eval_and_assim_can_match
which defaults to false for backwards compatibility
but can allow obs which are sucessfully assimilated in
one experiment and sucessfully evaluated in another to 
match.  the implementation is 0 and 1 match, and 2 and 3 match
(assuming the dart_qc_threshold is at the default value
of 3).  fixes issue DART-31.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@7002 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 2230e1979eb7ae90b42c1dc18fd5f9c9d9de8ddd