Now initializes the temporary U_wind component pieces to be missing. The U_flavor was uninitialized and could be checked if the FIRST wind component was a a V component. Now, it just indicates that there is a V without a companion U.
I put the 'missing u,v' print statements inside the verbose flag
so the warnings only get printed in verbose mode.

I printed the summary table of all the known observation types
and how many observations were possible - a single time, since
the number of observations possible is the same for the prior and
posterior.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@7850 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 68523ad0576b9d502b0399c2ffc1ef300cc72b0b