updates to the madis converters.  bug fix for the rawin converter - for height obs on significant levels it was incorrectly discarding obs.  it was also not using the right qc field for heights.
for the rest of the madis converters, sort the obs by time 
first before adding them to the output obs_seq file.
they run much faster this way, and almost all the converters
already read in all the data first before looping over it
so there's no increase in memory usage.

the updates to the path_names_* files are because they need
the sort module now.  also make obs_seq_to_netcdf which can 
be useful for diagnostics.

the "sig level by height" bug fix is already on the lanai 
release for the rawin converter.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@8762 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 9c5326fcc94b595bacdaf59251011d223398b364