Added the option to have POP variables be defined in the model_nml, rather than hard-coded into the model_mod.
Variables can be defined with 
  model_state_variables = 'netcdf_var1', 'dart_kind', 'update',
                          'netcdf_var2', 'dart_kind', 'update',
                          ...

If no model_state_variables list is provided, the default
(SALT_CUR, TEMP_CUR, UVEL_CUR, VVEL_CUR, PSURF_CUR) netcdf
variables will be used.

Diagnostic files only output the default variables.  This is
because there is no attribute information for additional variables
that could be added to the state.  This can be upgraded in the
future if there exists a need.

Currently the update option is not being used, but was put
in for convinience for when update information is added to
the state_structure.

Removed the print_range subroutine as it no longer is trivial 
to calculate min and max values with a distributed state.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@9162 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 15de0a9cf77ef9e2f89513fddd0dcccaa74aa113