merging trunk_with_cam_style back into the trunk.
Notes from Kevin:

This is expected to be the last commit of the old code style model_mod.f90.
The next version will have extensive style changes, but bit-for-bit agreement
with this one.

It has important fixes for errors in model_heights, etc., which were introduced
in the upgrade from using the nearest grid point for calculating heights on model
levels to interpolating to the ob location.  There was an extra division of
the surface geopotential by grav.  There was also an error in the conversion from
a 'do while' loop to a 'do ... ; if' structure when searching for the bounding
heights of an ob at a height.

These had no effect on assimilations of obs which were not on heights,
and at worst resulted in obs on heights being ignored when they shouldn't
have been.

Notes from Johnny:

Bitwise tested to the trunk with FV and SE with both vert_coord=pressure and log_invP.
Assimilated with GPSRO_REFRACTIVITY observations and used large observation
sequence files and full ensembles.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@9541 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 00e9cfda59d7cc33633c3949adf83d3e47b37062