This model_mod is functionally identical to the previous commit, but has been brought into line (mostly) with the DART style guide.
Most of the changes are white space, so turning that off in xxdiff
will clarify the real differences.  
Many changes are replacement of 'dimension(...) ::' attributes with 
':: variable(...)'.
Many changes are replacment of CamelBackStyleNames with under_score_style.
There are added comments for doxygen documentation.  They start with '!>'.

The style changes also revealed a mistake in previous model_mods.
The grid_1d_type P0 (0-D reference pressure in CAM) was using an 
incorrect dimension ID when being written from nc_write_model_atts.
This should have had no impact on any assimilations.  
Despite this, an FV test shows that the Posterior_Diag files are identical
when using the previous, non-style, model_mod, and this one.  Test results
are in case FV_new_damp_defaults/archive/dart/hist/Post_merge_{base,style},
and others generated by Hendricks.

This is the version which should be used as the basis for splitting 
models/cam into specialized versions.  The beautification work of the 
undertaker is done.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@9555 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 8f0dac660f85e7c57f1a41eaa06ed4dced68716f