Merging in branches/rma_diag. This contains dart_nc_write_model_vars and dart_nc_write_model_atts (state_space_diag_mod.f90). These where skeleton routines previously. These routines are used to write the STATE variables to Prior and Posterior Diagnostic files if desired by a model_mod. They are used in the same way as nc_write_model_atts and nc_write_model_vars, i.e. called in init_diag_output and in aoutput_diagnostics. Note the model_mod can write other info to the Prior and Posterior diagnostic files during nc_write_model_atts, e.g its input_nml or static data and still have DART write the state variables.
A diagnostic domain was added to the state stucture. This is so 
you can combine multiple domains into one diagnostic file. 
This domain is 'hidden' after the maximum number of domains so
it is not counted as part of the state, but all the domain, 
variable and dimension accessor functions can be used on it to
write netdcf files. The tests used to develop the diagnostic
domain are in io/tests.  There is an assert module that may be
of use in other tests.

Code corrections:
source, revision, revdate was missing from some E_ERR calls, 
these have been added.
Removed the optional argument from lorenz_96/model_mod
nc_write_model_vars. Never used this, it was left over from
refactoring IO. 



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@9746 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: dbb7e1b425053129116c99603cc86820bf191f48