fix a bug where the obs error field was too small to contain values > 10.  one of our users has a file from 2015 which has wind errors of > 10 for satellite winds.
changing the write format in the 2 prepbufr converters
requires a corresponding change in the read format in
real_obs_mod.

the "real" fix is to stop writing ascii intermediate files
and combine these two converters into a single program.
at the end of the prepbufr code we have all the bits we
need to create a dart observation, except the date.  (the
files contain hours only; the date comes from the filename.)

for now this allows users to convert these observations
without error.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@9759 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 306c33ed642cd7d0011b6256325fcb13a902b48c