The stream text files should have names like the defaults, that way they will not get confused with the ones that get staged automatically (and have the default name not be used ...)
The setup_pmo script actually is evolving to be able to do multiple instances
since Fei runs an ensemble for the free run. It can be modified to run
a single instance, so why not.

Right now, the resulting job has not been run successfully, I was staging
a CICE4 restart file for a CICE 5 run ... need an ensemble.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@10397 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 88455b256501de75303cb9ec55088007e2ed70c4