in the utilities module, update the open_file() code to support more options. we can now use this single routine instead of custom open() calls in other modules. make the error messages more specific (hopefully now more helpful). make it possible to return an error back to the caller for handling. the defaults have not changed, and without an optional argument this routine still fails here on error.
the obs_sequence_mod now uses the updated open_file code, and does
not try to handle "pre_I" files anymore (which can't be supported
anyway with our current obs kind/types code).  also added a namelist
option to read binary obs_seq files written on a different endian machine.

updated the cice converter to use the open_file() code with the
stream input format option, and fixed up the defaults in the example
namelist so it runs with the single example obs file we supply.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@10622 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 5637abf04e44ae90f7cfdf245efdfddb71b52d6d