No longer using the obs_type variable as the basis for interpreting the obs_provenance strings. This is at Andy's request and OK'd by Chris.
I also made it an error to use the 'IJK' locations since this is not working 
correctly, as far as I can tell. When converting the observations using IJK,
one does not get the same locations as when the lat,lon,depth locations
are used that come native from ROMS. The default was to not use the IJK locations,
and I removed the namelist variable that would enable it. That variable had already
been removed from the models/ROMS/work namelist and examples that Andy has
been using - so there is no expected impact.

Updated documentation and example namelists.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@10865 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 6f118ff3253f4b0f42444a16631365efc4a74bf2