There are variable attributes that help understand the context of the variables. The dimension and variable containing the list of ensemble sizes supported is now called 'ens_sizes'.
Minor cleanup on formatting for the rest of the source codes.

Changed the html document to reflect the intended usage of the 
sampling_error_correction_table.nc lookup table.

Note that the actual values in the lookup table are a little different
than the values in the final_full.N files, and that the assim_tools
code does not actually use the netCDF lookup table yet.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@10961 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 4d5c283e6230ac33f8f8db859fe6183c32a153f5