This uses the additions for the POP2 namelists without breaking the earlier pop_in namelist support. Also supports along-track sea level anomaly observation assimilation.  The long-term mean dynamic topography file is read from the &model_nml:mdt_reference_file_name variable.  The size of the 'mdt' variable is compared to the expected size, the units are converted to SI upon read - as the observations are also SI, and the forward operator requires SI.
The setup and assimilate scripts work with the new assumed filenames
and support the netCDF-only logic.  The assimilate script allows
DART to modify the restart files directly - no safety or backup files are created.
If the assimilation corrupts the POP .r. file, you must revert to a previous
restart set.

The AVISO converter 'makedaily.sh' script is simpler and supports the
addition of the WOD observations into the output observation sequence file.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11279 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: afa936f2638fd5c2a03858d10c91792b0a4aca7d