Move filter.nml to modules/assimilation/filter_mod.nml, so that it's in the same directory as the module which defines it.
Split the previous programs/filter/filter.html into a new version,
which focuses on the filter overview, and modules/assimilation/filter_mod.html,
which focuses on the filter_mod_nml.
Also general updating of the content to be consistent with the Manhattan release -
fixed filenames, 4 state space output stages, etc. - and condensing and clarifying
some text.

The htmls would like to be reviewed by someone else's eyes.

NOT done; 
   > updating the ERROR codes sections of the html files.
   > testing all the links.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11406 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: d283c7f5b0c78dacef03af11c8d25d0d8f5e4067