This is for an intermediate CESM, before the (hopefully stable) release in June 2017. It was used for Bretherton's clouds experiment and for interfacing the developing Manhattan release and CESM2. It's based on the stages {in,out}put, {pre,post}assim, but no {forecast,analysis}, and doesn't include Nancy's model_mod changes of ~2017-5-11, or changes for running on Cheyenne.  It is a step in that direction.
There are many fixes to the st_archive tool, which hopefully will disappear
from DART when CESM2 is released with a st_archive which handles DART output.
Config_archive.xml is a replacement for the CESM version because the original
uses the wrong file name form for creating rpointer files in archive/rest/$date.
Case_st_archive.py is also a replacement due to many multi-instance errors
in the original.
There is also a section in setup* which modifies the env_archive.xml to handle
the current stages of DART output.  

The disk space management by assimilate.csh during runs has become complicated
with the stages in Manhattan.  I'm considering making it a separate script,
called by assimilate.csh.

Setup_* (test versions) and assimilate.csh were tested in cases
  setup_advanced: CAM6_test1_adv
  setup_hybrid  : CAM4_test1_hyb
which are 3 member, single day assimilations.

The 2 spinup_single* scripts should be condensed into 1.
For the moment it's useful to leave them separate, since that's what
was used for the development work of this CESM.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11634 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 481ccec9c136c5494617be9625e279062f995ff9