These functions use whatever filename is specified by 'diagn_file' as the default file name to use when prompting. Whatever you specify at the prompt is used, but does not affect the setting of 'diagn_file'.
You can specify the true state via  the 'truth_file' variable,
which has a default value of 'true_state.nc'.  This is not prompted for,
by design.

There was an error in CheckModelCompatibility in that the model name
in each of the files came from the same file - so they always matched. Oops.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11649 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 0fa3ac5907ffcafb01f53aba2f4ca54bf1468b34