add lorenz 96 2-scale model mod to the rma release. updated diagnostics to match.  added 2 new obs types and quantities for this model (large scale and small scale).  it was only working before because it was only assimilating identity obs.
update the test routines to allow '-default' as a flag
which says compile the models with or without mpi depending
on what the default for that quickbuild is.

add new routines to the netcdf utils to create more types of
variables and attributes.  add more options to allow the
location variable to have a name different than 'locations'.

change failed vertical convert from QC 7 to QC 4 - after testing
it should be changed to QC 8.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11683 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 96353d76a1e5b9735f1de5eef0bd172e2f6a5a4a