restore a "working" input.nml.  the previous version was accidently overwritten by the workshop version and committed.
NOTE that this one differs from what we distributed with Lanai,
but it is simplified and gives a good assimilation result.
it turns off inflation and turns off localization.  this is
intentional.  the goal of the default input.nml is that if someone
just runs filter, they should see a reasonable assimilation result
and not write us for support.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11688 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: b248b503647b7b9ea83ddf8315f93b040f9619a4