Merging in the branch rma_mmc.  The code can now run with oned, twod, threed_cartesian, and threed_sphere location modules.
There is additional code to run model_mod_check with
the single_file option (all members in one file) in
the namelist 'single_file'.

restart_file_list has been renamed to input_state_files
and similarly for output_state_files.  

Moved model_mod_check to the programs directory and updated
the html and all model's path_names and nmls to be consistent 
with the location of model_mod_check.f90.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11752 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 958d6936d8f9055fcd79f189b8425fdd0428e369