fix an 'off by one' error in computing the specific obs types which correspond to a generic observation quantity, when building an obs impact table.
fix some code that validates that there are enough members in
a single-input file - error checking was too strict if starting 
from a single member and perturbing it.  also rearranged some
code that was calling a validation routine once per loop instead
of once up front (did not depend on anything inside the loop).



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11817 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 38dc2b884496e596ee796f664af086f604136f26