merging the rma_fix_clm_restart branch into the trunk.
clm_restart files are now bitwise with the trunk, with
the exception of H2OSNO, which has some slight
differences which has to do with the order which 
clamping is applied.

adding clm_to_dart and dart_to_clm, which marks missing_r8
values to the state.  There is one additional namelist for
filter_nml 'allow_missing_clm' which is only to be used for 
CLM assimilation. this was previously in assim_tools_nml

there is an additional module options_mod.f90 which gets and 
sets whether or not you can have missing values in your state
vector.  this avoids circular dependencies and can be accessed
directly by the model_mod.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11920 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 28860027ff66486606cd97eed0a8c4d0bccb7247