This data file was created by I-Te Lee: "First of all, the original idea  error of ionospheric observation is 1%. Thus, I put the code as "oerr = 0.01_r8 * obsval". Liu et. al and Yue et al investigated the Abel inversion error of COSMIC ionosphere profile, both of them figure out the large error would appear at the lower altitude and push model toward wrong direction at the lower ionosphere while assimilating these profiles.
On the other hand, the Abel inversion error is depended on the ionospheric electron 
density structure, which function of local time, altitude and geomagnetic latitude. 
Therefore, to simplify the procedure to define observation error of profiles, 
Xinan Yue help me to estimated a error matrix and saved in the file which named 'f3coerr.dat'. 

The file included a 15 by 37 by 25 (altitude - latitude - local time) matrix, 
ranging from 100-800 km within 50 km interval, -90 to 90 degree in geomagnetic 
latitude within 5 degree interval, and 0 to 24 local hr with 1 hr interval. 

The number in the matrix is error percentage (%), which calculated by OSSE.  
Please check the attached files for that two reference papers .

In the end, the observation error is summed up of instrumentation error (10%) and Abel error."

I converted the file to netCDF, the references are included as global metadata.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11930 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: aa57533ac120a2f0435a2b0ab81bcf8d525e9b41