trans_coamps_to_dart  works and uses the hdf variable names for the variables in the hdf files. The nuisance variables have names that are consistent with the naming scheme. The hdf_names are now part of the state variable structure and are filled during the 'coamps_[read,write]_all_fields()' call.
Unfortunately, those routines are not called by static_init_model().
There are two instances of the state_vector - one for the DART state, 
and one for the COAMPS state. The list of 'variables' needed to construct
the DART state vector are contained in 'state.vars' but this table only has
the short name of the variable (UUWIND) and a list of attributes.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/coamps@11934 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: ccb886f24ee77c97439cc5a40a0abee6e9b7fa1b