Getting closer.
with the default 1 observation on VERTISSURFACE, generates an error
 ERROR FROM:
  routine: deallocate_available_values
  message: Deallocation Error::available_target_values
  source file: $URL: https://svn-dares-dart.cgd.ucar.edu/DART/branches/coamps/models/coamps_nest/coamps_interp_mod.f90 $

with two observations (one an identity ob, one on vertislevel) it runs but gives bad values for obs_seq.out

I am having trouble updating an HDF5 file with the netCDF 4.4.1 fortran modules.
Supposedly using 'nccopy -k 4 coamps.hdf5  coamps_updateable.hdf5' results in an HDF5 file you
can modify with the netcdf module ... hasn't been my experience so far - the variable in 
question EXISTS in the 'coamps_updateable.hdf5'.

 ERROR FROM:
  routine: nc_put_real_3d
  message: put values for uuwind_sig_028030_000010_1a0201x0204_2013011000_00000000_analfld: NetCDF: HDF error
  message: ... copy_netCDF_to_hdf
  source file: $URL: https://svn-dares-dart.cgd.ucar.edu/DART/branches/coamps/assimilation_code/modules/utilities/netcdf_utilities_mod.f90 $

And even then, I'm not sure COAMPS would be able read it back in.

Given the way the rest of coamps seems to be written, there is probably a coamps_hdf_mod.f90  that we
might be able to use.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/coamps@11938 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 71b751a48dc13962b054bb01e43773265258da89