bug fix:
don't call convert_vertical_obs() on tasks which don't have obs.
model_mod code might not be expecting a size of 0 (and better not
to assume it would.)  do the same with state even though currently
we don't allow #tasks to be larger than the state vector size.

fixes for other things:

- increase integer format width in count_close messages
- add timer around count_close() based on brian's comment that
   the all_reduce is slow on large task counts
- combine 2 missing_in_state vars into one
- add comment in spread_restoration about the fix craig proposed
   for the case of 0 spread in the obs prior variance.  not part
   of the code yet because it might not be sufficient
- remove old debug section in code that "can't happen" to avoid
   confusion



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11950 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 1d423f8f5c1ec1aac2a1b4cc519bac08bc50a848