in the dart_obs_seq_mod, added a width to a write format spec because gfortran forces 'i' formats to have a width.
added the -D_REAL8_ cpp flag in the mkmf_gsi_to_dart
makefile.  this now successfully compiles on cheyenne
with both intel and gfortran.  i left comments about
the older way of building this executable in which
it used cpp to preprocess the file.  on cheyenne with
gfortran this didn't work because c-style comments
were inserted at the top of the file which caused the
fortran compile to fail.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11991 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: da919065c88f442f1095f441310c1d2d3a6d6d29