add RAW_STATE_VARIABLE to the simple advection defined observation types since it doesn't use the standard obs_def_1d_mod.f90 file.
minor fixes in the obs_utilities to handle the cases where the
caller asks for the fill value and none exists for that variable.
it should return MISSING_R8 in that case but in some cases it
returned unset and in others it was a fatal error.

update the change log for last few changes.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@12136 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 7512db12590cb357e1ea97fc1586a8c8fb7f6e9d