move updates from rma_trunk to Manhattan release. changes include:
- Enhanced inflation algorithm added.  See the filter_mod.html for new
documentation on this option.

- Updated WRF scripts for the Manhattan release.

- Obs_diag reports statistics on DART QC 8, observation failed vertical
conversion.  Matlab scripts also updated to support QC 8.

- New parallel conversion scripts for GPS Radio Occultation observations and
  NCEP prepbufr conversions.

- Further updates to documentation files to change KIND to QTY or Quantity.

- Documented required changes when moving from the Lanai/Classic release to
Manhattan in documentation/html/Manhattan_diffs_from_Lanai.html

- Expanded the routines in the netcdf_utilities_mod.f90

- Add an ensemble handle parameter to the 6 ensemble manager routines
where it was missing.

- The advance_time program can read/generate CESM format time strings
(YYYY-MM-DD-SSSSS).

- Fixed a bug in the netcdf read routines that under certain circumstances
could report an array was using the unlimited dimension incorrectly.

- Removed the option to try to bitwise reproduce Lanai results; due to the
number of changes this is no longer possible.

- Minor bug fixes to the (seldom used) perturb routines in the WRF
and MPAS_ATM model_mod.f90 files.  (used to add gaussian noise to a
single model state to generate an ensemble; this is never the recommended
method of starting a new experiment but the code remains for testing
purposes.)

- Several remaining model-specific model_mod_check programs were
removed in favor of a single common program source file.

- Keep filter_mod.dopplerfold.f90 in sync with filter_mod.f90,
and assim_tools_mod.pf.f90 in sync with assim_tools_mod.f90.

- Removed makefiles for the obsolete trans_time program.




git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/releases/Manhattan@12591 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 6024b2e813bc139b30d801f11bb18422293affc5