interface to cloud cleared level 2 radiances - granule info from example data reader.  obs converter now making obs_seq files  with metadata for satellite/solar location, and platform/sat_id/sensor/channel.  can read/write them now. created a test input obs_seq from a downloaded CC_RAD hdf file.
forward operator code is building temp/pres/moisture profiles
on model levels.

now trying to fold the example program code into an rttov_interface
module so it can be called by the forward operator and keep the
rttov calls isolated as much as possible.  still lots of questions
about surface fields and some of the other inputs the rt model
code needs.  this file needs to be preprocessed.  that's tricky
on the macs where the filesystems ignore case (which is insane).
so bob.f90 and bob.F90 can't be distinguished.

these .f90 filenames need to be revamped eventually to be clear.
we still have the existing airs retrieval converter and right
now the names are confusing.  do this later when more code is
working.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/radiance@12747 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 6b8f333b97e271bfa0148efea36dd239d543ff51