add code from cesm branch that supports an external communicator passed in at init time.
consolidate the calls to the system() subroutine
since the form varies by compiler. remove replicated
code by consolidating where the optional initial string
(usually for a shell), and a trailing 0 for c-style 
termination is added. that code now occurs in a single place.

move all the mpi_reduce-related calls so they are
grouped together in a single code section.

doxygen-ize the comments.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@13119 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 8a5b13a97675c6516841b6e95af9ff92ce51e12a