A script to repackage CAM6+DART output into forms that are useful and efficient for storage and future use. 2 sections have been developed and work in the testing context: "forcing" files in the form of coupler history files, and restart data sets, including inflation.
The restart section will call mv_to_campaign.csh, 
which submits a request to globus to transfer files,
but that has not been tested in this version.
Instead it writes out the mv_to_campaign.csh command
that would/could be used.

This script should be run as a batch job, 
since it uses the cmdfile mechanism.
If it calls mv_to_campaign.csh, it must be run on casper,
since the cheyenne compute nodes do not have access to globus.

2 more sections will be added; obs space diagnostics
(obs_diag and matlab scripts must be run first),
and DART diagnostic files (esp/hist/..., atm/hist/$inst.e.$stages).

It has not been reviewed.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/reanalysis@13135 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 7676d6a3c77e3a834a1a2afc68fcf03bc4a9ba4b