really minor updates to the _quantities.f90 files; only a change in the comments to be less confusing.
more substantial but relatively minor changes to the
preprocess.f90 source.  changed 'kind' to 'qty' in 
several more places, some in code, some in comments.
pulled more replicated code into subroutines.

committed to the preprocess branch since this is
working as-is. then will start adding code to support
adding units and bounds to the quantity files.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/preprocess@13186 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: abc4e0e49dd8a515acbea530595b86d93515f92f