kevin discovered that updated netcdf software on cheyenne will no longer open our inflation files (in filter) nor our diagnostic files (e.g. ncview, obs_diag?) because we've given the 'lev' and 'ilev' netcdf variables a unit of 'level' when it's really 'hPa'.  udunits throws an error and the program exits.  i haven't found a way to avoid this.
in the meantime, this commit adds kevin's changes to the
units attribute in the cam-fv/model_mod.f90 (changing 
'level' to 'hPa') and that should fix new files we create.

i'm not sure how we deal with older diagnostic and inflation 
files going forward.  perhaps using 'ncks' to alter that
attribute in the files?

i checked the initial and restart files CAM writes out
and they have 'hPa' as the units.  we didn't notice this
when updating the cam model_mod.f90 (where the variable
attributes are created by our code).



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@13187 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: a9d6581fbc0041099a247594d77aa0cf1c33af72