make a new branch called anamorphosis off the rma_trunk so moha and i can work on the combined changes needed to support it.
most of the work is from moha and in either assim_tools or
filter. but i'm taking this as an opportunity to revamp preprocess 
to make the quantity list also autogenerate, and to require units 
as well as a quantity name.

i will be merging changes from the 'preprocess' branch to this
one soon.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/anamorphosis@13195 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: e6ea38d10538556530227d53cdae3cbb15564e54