start of making gitm work with manhattan.  NOT FINISHED!!
created a new executable to convert a single gitm .bin
file to a netcdf file (instead of a lanai-style restart file).
it's unclear to me if what i have is a single block of a larger
dataset (since it has ghost zones) or if this is the full dataset:w

modified all the path_names files to compile with manhattan.

started work on the model_mod - made the interfaces work.
this will NOT work for filter because get_close_obs() will
need fixing, and there isn't currently a get_close_state().
also may need to write the 2 new vertical convert routines.
for now i'm just passing through the default versions which
satisfy the compiler but do nothing.

also made the quickbuild match what we currently use in
other models (doesn't build mpi targets twice, etc)



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@13208 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 00060bcf1a0d26bc4d0384492bbb599bb8cc20fb