fill inflation bug fix; init/final calls added to wrf utils; doc added.
fixed bug in the fill_inflation_restart program. it was hardcoded to use
the prior inflation mean/sd values for both the prior and posterior inflation
files.  also fixed documentation where it says that it ignores the input
file list.  it is used as a template file for the sizes of the variables.

fixed a simple syntax error in the location test driver script.

added some info before each subroutine in the default_model_mod.f90
file with hints about what each routine needed to do.  the full docs
are in the default_model_mod.html file, but if you're looking at the
source code this reminds you about the point of the routine.

several of the utilities in the WRF_DART_utilities directory were
not calling initialize_utilities() or finalize_utilities().  we've
made that more of a requirement.  in past versions of the system if
you used another utility routine it would call init for you - although
no one called finalize for you.  but that could create recursion loops
in the init code, if it was called from a routine that was then called
by the init code.  it's just better if programs explicitly call the
init and final routines for the dart utilities.