rename gitm converter files to be more accurate for the manhattan version.
in the manhattan version we're converting between gitm blocks or a single
gitm file and netcdf, not a 1D dart state vector.  change the names so
at least for now they're less confusing.

we're still missing a netcdf_to_gitm.f90 which writes out just the
model state in a single data block - but i'm not sure what to do with
the ghost zones so for now i'm waiting to see if/when it will be needed.