make the test script continue even for programs that read standard input.
several of the converters prompt the user for input.  change the test
script so it redirects stdin to /dev/null - it will try to read that,
get an error, and continue.  otherwise it simply waits for input from
the terminal, until you kill it.

to give input to converters that need it, make a file of the form
<convertername>.in and it will read input from that file.