fix location and name of hardcoded input file
the name of the input file has apparently changed and it's now
in the ../data directory instead of the work dir.  fix the input
filename variable in the code so it just works after you build it.

i guess at some point the program could include a simple namelist
which contains the input and output filenames as an example of
how to make it possible to set the input name at runtime.