fix a place i still had 'tests' instead of 'work' for input files.
if you untar the test file bundle, it adds an advance_time.in file
to each of the observation converter work dirs.  i'd toyed with
making a ../tests directory but that seems unnecessary so i removed it.
but i missed one place in the test script that still looked for it.
this was causing all the tests to fail running advance_time.