bug fix for vertical conversion code.
was not setting the return code correctly when converting
the incoming obs vertical to the localization vertical.
this usually doesn't happen if you convert all your obs
up front, but if the obs vertical type is VERTISUNDEF it
isn't converted and it gets into this code.  the return
status is unset and you get whatever random trash is in
memory at that loc, which is frequently not 0.