query for and take whichever significant wind obs are present in the file.
there was an error in the logic if you had asked for sig winds but
didn't have a second logical set to pressure or height correctly.
these files seem to only have either height or pressure for sig winds,
not both.  query for whatever is there and use it.

also cleaned up some of the code that was a little confusing,
added some comments, and added a 'verbose' option to print out
a little bit more info about what was found in the file.