fix remaining bug in madis rawin converter; add init to cice converter
the madis files have either 9999 or 99999 as the missing value
but no missing value attribute.  i clamped the max valid count
at 1000, but didn't reset the count to 0 if that was the case.
this commit fixes that.

also, the cice_to_dart converter wasn't calling initialize_utilities()
or finalize_utilities().  it does now.