saving test logs for all output
The bgrid_solo model code _may_ require the compiler directive -Duse_netCDF
but only for the programs that use the fms_src. The compiler directive
is not needed right now, but is uniformly implemented across the appropriate
routines.

The var input.nml did not have namelists for some of the programs tested,
so it was never getting off the ground for the test_dart.csh script.