Preserving files that were under review 2019-8.
Some of the files being committed were in the process of being updated
based on a code review in 2019-8, so they haven't been tested in their current form.
The changes that were made were not propagated back to the svn directory,
so the svn and git versions of some files have diverged, and will need to be
merged after each of those sets is committed to git.
The main changes are in repack_st_arch.csh.
They're mostly about changing variable names to be more descriptive
and easier to use, improving comments and echoed information.
There are changes to functionality to improve efficiency ("Previous" directories).