Generic version replaced by Reanalysis version.
This branch ('reanalysis') is specifically for that project,
and does not need to be easily usable by a new DART+CAM user.
And now we will be archiving the version of setup_advanced
used to build a CASE in the github repository DART_CASES,
so those versions should not live here as new files.