Failure to modify mkmf.template dies nicely
In light of certain users simply attempting to run DART with the default
mkmf.template, then when it doesn't work say "it's broken," the decision
was made to remove the mkmf.template from source control. The user must
now select an appropriate mkmf.template before proceeding. This makes it
a bit harder to get started, but the difficulty is necessary so that
users will be able to handle issues such as not finding NetCDF.