Adding developer tests for RTTOV forward operator
rttov_test.f90 creates a sample rttov_sensor_db.csv, populates data
according to some profiles that were distributed with RTTOV, then
forward calculates both the all-sky AIRS IR and AMSU/A MW radiances.
This is a fairly basic developer test, but it should hit most of the
important branches of code.

Note that the developer test does not include aerosols or the like.