added info on what types of seeds we use in various places in the code
tim converted the standard html docs to markdown.  the process is not
very good - these files have the content but need serious reformatting.
not complicated but tedious.  probably a job for a python program or
some fancy sed scripts.

as far as content, i added a section at the top of the random_seq_mod
explaining the various types of seeds we use in the code and why.
in the time_manager_mod i added a block for the generate_seed() routine
which was missing.  it takes a time type and turns it into as unique
an integer, appropriate for a seed value, as possible.