Adding absolute humidity observation converter
necessitated changing the observations.html document, so I converted it to README.md
so it would be more 'viewable' on GitHub.
necessitated adding something to the CHANGELOG, so I converted it to CHANGELOG.md
in the hopes it will be 'prettier'. No content changes, but I did notice that
the support for DART QC 8 got added on several occasions. not great.