Adding CISL-suggested directive for real-time log output.
"The recommendation is to add directive #PBS -k eod to scripts to reduce the
risk of job failures that kill Cheyenne nodes. With the change, the job will
write output directly to its final destination directory as it progresses
rather than only when the job ends. This will prevent the node from being
killed if the user's application writes too much data, which is a common
cause of Cheyenne node failures. In addition, if a node dies at the end of
the job, using this directive will increase the likelihood that error messages
will be preserved."