bugfixing perturb routine to generate ensemble that can have missing values
Very few models have missing values in their state. There was
a bug in the routine that created a new ensemble from a single
member. The locations of the missing values was not being handled
properly. Only CLM could have used this routine, but since CLM does
not allow perturbing from a single instance for production, this
bug could only affect tests.