small responses to pr comments
Commenting out the NETCDF, HDFRIR and RTTOV variables
so that if they are defined in the user environment they
do not need to be specified in the template. If they are not
defined in the user environment, the script fails and there
is a gigantic comment block explaining what to do - inside the script.