Removing untested converters for an unwanted data stream.
The idea was to convert a "L2 cloud-free radiance" product from AIRS,
which means it used AMSU to help clear the clouds and/or fill in the gaps
with a retrieval of what AIRS "would have seen" if there weren't thin clouds.
The README for that product says most users probably won't be interested in this
product, but NASA provides it because it is included because it is an input to
other L3 retrieval products.

https://catalog.data.gov/dataset/airs-aqua-l2-cloud-cleared-infrared-radiances-airs-only-v006-airs2ccf-at-ges-disc
"many Standard Product users are expected to have little interest in the Cloud Cleared Radiance"