Works with r8 and r4, ens_size >= 2
An ensemble size of 1 will cause an MPI error.
The CPP preprocessing is now explicitly performed. The output file name
is intented to be slightly different than the input name to accomodate
case-insensitive filesystems. The endian-ness of the files is explicitly
determined in the open_file() command as opposed to a compiler directive.