ensure backwards compatiblity in namelist
add the eventual new items to the namelist, but initialize them with the
existing namelist items if they aren't set by the user.  also allow the
string 'obs_type_definitions' at the start of an obs_seq table of contents
instead of 'obs_kind_definitions' which is misleading.  either are accepted
so backwards compatibility is maintained.