fix the way i tried to rename these modules.
i guess if you do a 'git mv' and then don't do a commit first,
if you also update the files it can't tell anymore they were renamed.
i tried to back that out but unsuccessfully.  so remove the older
files.  i'm not sure what's going to happen to the history.