Removing io routines no longer used.
The 'sv_to_pv', 'pv_to_sv', and 'snapshot' io support is not
used in the Manhattan version. Adding a prototype of
the pert_model_copies() routine that needs to be extended.
The model can have 0.0 values in the state to indicate dry land,
these should not be modified by the pert routine.