Modifying declaration of hexadecimal constants
The latest mp-gcc10 compiler does not support declaring a 'z' construct without
it being a data statement or an argument to an intrinsic function; e.g., int.
The hexadecimal constants are now declared
as arguments of the intrinsic int function. With this modification, the code now
compiles correctly using mp-gcc10 and also earlier versions (such as mp-gcc9).