attempt to make a .nml which is more helpful and correct
the nml file is tricky for this program because i've added code to
support older format namelists as well as the new namelist we want
to convert over to.  the code has a place where at some future point
someone can insert a nagging message saying older items are deprecated
and what they should be replaced with.  for now, this .nml file lists
the new items, and then has some "comment" lines (outside of the &start
and / ending lines) which indicate where to find appropriate values,
and which namelist items are to be deprecated.