updating with many quantities from public master
The branch used for this development work did not have the latest
quantities.  I have only updated the all_quantities_mod.f90 with
the minimal information ... it will be VERY useful to go back
and add the 'desc' to many of the quantities - which I have not done
yet. Nor have I updated the logical subsetted quanties_mod.f90 files.