exploit environment variable 'DART' to create filename.
The Fortran 2003 standard has an intrinsic get_environment_variable()
that allows us to correctly set the path to all_quantities_mod.f90
instead of trying to set a default value with a relative path. A single
value cannot work because preprocess gets called from different 'depths'
in the DART tree.