reproduces output from NASA example program.
The BUILDME script can build the hdf-eos libraries such that
you can use an external netCDF library - which will be necessary shortly.
If the HDF-EOS5 sources are available, those get built as well.
HDF-EOS and HDF-EOS5 are installed in separate locations, which seems
to be standard practice. The download instructions and compiler settings
are not finished.