Adding Obs Network choice functionality to L96_inf
The changes are summarized below

(i) The user can now select between 6 different obs networks:
- Full: Observe all 40 variables
- Half: Observe every other variable
- Quarter: Observer every 4th variable
- 1:20: Observe the first 20 variables
- 10:30: Observe the middle 20 variables
- 1:10; 30:40: Observe the first and last 0 variables

This should allow the users to understand the interaction between localization,
inflation and the density of the observations. A lot can be learnt from this!

(ii) Improved the functionality of the polar plot:
- The plot is now annotated from 0 to 40 rather than 0 to 360 degrees
- Removed useless annotation in the center of the plot
- The xlabel on the inflation plot is edited accordingly
- The position of the figures have been adjusted to fit the new obs functionality

(iii) Variables in the code have been edited to mimic the fortran code
- Inflation variables inside 'run_lorenz_96_inf.m' and 'update_inflate.m'
have been renamed such that they match those in 'assim_tools_mod.f90' and
'adaptive_inflate_mod.f90'