Update POP user_datm.streams & scripts for cesm2_1
Since the coupler history files for the CAM Reanalysis are now
unzipped on RDA, we edit the POP user_datm.streams* templates and the
CESM2_1 setup scripts for running PMO and setting up a hybrid ensemble
to enable them to make use of the unzipped files.
This fixes Github NCAR/DART Issue #87.