Full HDF4 support and build instructions.
The BUILD_HDF-EOS.sh script builds HDF4, HDF5, HDF-EOS and HDF-EOS5
libraries and dependencies. There are instructions on where to
get the sources.

The airs_JPL_mod.f90 was originally written for HDF-EOS and I
naievely updated it to use HDF-EOS5 thinking we would use it
when converting radiances (the RTTOV dependencies are for HDF5).
Since that's not happenning, I am reverting convert_airs_L2.f90
back to use the HDF-EOS (HDF4) routines as originally written.