AMSUA converter working, still need help from Jeff S
The machinery is working, but I need help with the
add_granule_observations() routine, specifically the part
"      ! calculate the bearing between the obs lat/lon and the SClat/lon"
Specularity, azimuths, that sort of thing.
There are some lines commented out that say 'HELP'.
Those are the areas with questions.
NOTE: I am not checking error codes or implementing any kind
of guidance from the documentation - yet.