Simplified HDF-EOS, HDF-EOS5 configuration.
There is no reliance on using the Cheyenne module for the HDF-EOS5
support and the mkmf.template.rttov.* files are functionally very
similar.  The mkmf.template.rttov.intel still works if you DO have
the hdf-eos5 module loaded - although if the version changes, the
value of HDFEOS5 should probably change.

The quikscat observation converter also used HDF, so the same
strategy is used to build it as the AIRS converters.

Since we are specifying the library paths in the mkmf_ scripts,
there is no guarantee that the run-time environment will be consistent,
so I made comments to that effect in the mkmf_ scripts and modified
the run_tests.csh to point this out.