update obs_kind_mod.html
“Knowing the QTY_xxx of an observation must be sufficient to compute the correct forward operator.”
I don’t think this is true because the case statements for which forward operator to call are by type. 
This has been in the docs for a while. I think this is an overzealous find and replace from 
commit: 7e2687b19b10020e0e03774c5204e2b97e2eb916