Update preprocess.f90
Changed the comment to match the code.  If {input/output}_obs_qty_mod is not set, input_obs_qty_mod_file = input_obs_kind_mod_file
so the defaults are  DEFAULT_obs_kind_mod.F90 and obs_kind_mod.f90.

I've left in the todo comment because it still is a todo to change all the use statements, filenames and pathnames.