rename new routine to get_value_from_string()
make the name still general (could be extended to return other values)
but a little more of a clue about what it is doing.  update both
filter_mods to use the revised name.

add some notes to the utilities module about 4 subroutines being located
in the wrong section of that file.  i'll open an issue about moving them.