add background color for table header and odd rows
There are extra wide tables on the following pages:

 observations/obs_converters/Ameriflux/level4_to_obs.html
 observations/obs_converters/ROMS/ROMS.html
 observations/obs_converters/SST/SST.html
 theory/readme.html
 models/tiegcm/readme.html

These tables go over the edge of the text area.
The coloring of the table rows, vs transparent is to make these
extra wide tables not show the ncar waves image behind the table.

In the future we may want a better solution for these super wide
tables:

* not use tables
* have wy-nav-content {max-width: none;}  so people can resize the
  text to the width they want by resizing the window.
* Have more responive tables within readthedocs.  It is not clear
  how responisive readthedocs is designed to be, for example when
  clicking on the menu in mobile mode.