added 4 new guide files, need to be added to a TOC if people like them
advice_for_new_collaborators:
i wrote up some of the topics we typically cover when someone
comes to visit us for a possible collaboration in porting over
a new model.  i think it's reasonably complete.

programs_that_come_with_dart:
apologies if this already exists somewhere.  i looked in the guide
directory and didn't find it.  it's a brief list of the programs that
come with dart, with links to the specific rst files, hopefully.

data_management_issues:
this is less fleshed out, but i think it's missing and needs to be
in our docs somewhere.  how we manage moving data between MPI tasks
is complicated.  people trying to add code to the system need to know
what they've got access to on their own tasks, whether they can get
data from other tasks (e.g. a window exists so get_state() can be used),
or if the data is replicated.  this doc is more in an outline form
but i'd like to add it even so and add more to it later.  i just don't
want it to get forgotten.

vertical_conversion_issues:
again, this may be someplace already but i wanted to list some of
the considerations when supporting models with different vertical
coordinates, including models where interpolation locations might
depend on the individual ensemble member values, etc.