make setting the error variance more explicit
there is a continuum of code styles from "as terse as possible,
no extraneous lines or variables" to "chatty, local variables
everywhere, spread the code logic out".

performance-critical code that doesn't change much leans towards
the terse side.  examples and tutorial codes lean towards
the verbose side.  terse code can be difficult to modify if needed,
but chatty code can make it harder to trace the logic if enough
unnecessary lines are added.  most code should be somewhere
in the middle.

this code was firmly on the terse side because i wasn't intending
to have it be an example for anyone and when i learned to code
it was a virtue to use as few lines as possible.  but if this
program turns out to be useful for users to modify for their own
uses it should be a bit less cryptic.

i've added a couple local variables (which could have been the same
o_err variable, but hey if i'm stepping into the chatty side of the
room i might go all the way) to make the error setting code a bit more clear.

[thus ends today's treatise on coding philosophy 101]