separation of FillValue and missing_value
If a variable has both a _FillValue attribute and a missing_value attribute,
make sure the values are identical or we cannot support it with a single DART missing value.
Handle _FillValue and missing_value separately because variables may have one or the other but not both. Adding a couple trivial unrelated changes not worthy of their own pull request,
removing unused variables, making a hardcoded filename a parameter, that sort of thing.