Removing complicating missing_value, _FillValue logic
The ultimate goal for models that support missing values is to read
the variables and replace both _FillValue and missing_value items
with the DART missing value. As long as r8 /= r4, there is no loss of
precision when writing BACK to the original files. If DART is built
with r8 = r4, and it reads from a NF90_DOUBLE and writes to a NF90_DOUBLE,
there may be trouble ...