clm_to_dart replaces bogus values with _FillValue
This is a standalone program and replaces bogus values in layers of snow
with the _FillValue value. It replaces the values in-place - i.e. it
modifies the input file. To use it, you should COPY the clm restart file
to the desired input filename. That filename should be the source of the
variables used in pmo or filter or ...