Debug version
Explores what happens to related snow variables when there is 'no snow'
according to SNLSNO. T_SOISNO in the layer closest to the ground still
has values even if there is no trace of snow (H2OSNO = 0.0). This seems
like a problem with CLM logic. frac_sno seems like a reliable indicator
to prevent replacing values in the layer thought to be 'empty'.