fixed localization issue for vars on 'lat,lon'
The 'related()' function did not support variables dimensioned with 'lat','lon'
and made them FAR away. This is the opposite of what is supposed to happen
for these variables. Anything aggregated to the gridcell level SHOULD
use the traditional distance-based cutoff.

The problem arose when I tried to assimilate a SOIL_TEMPERATURE and had
TSOI(time,levgrnd,lat,lon) from an XY history file to provide the
QTY_SOIL_TEMPERATURE. The forward operator is correct, but the posterior was
unchanged.