startup freerun builds ... note on startdate:
It appears that the 'startdate' has to be the first of the year.
I had it as 2010 07 01 and got an error/warning from buildnml
complaining about the clm namlelist using 'ignore_ic_date' not
being allowed.  I am not explicitly calling 'ignore_ic_date'.
The message went away when I started from the first of the year.