using DART indexing routines
There are assumptions about storage order that still need to
be pruned out.  the get_val() function in particular. Looks as
though there is an implicit assumption about depth being the slowest
moving dimension - given that, 'offsets' relative to the variable
start index are calculated and everything is treated as a horizontal
slab. There are native DART routines to calculate the explicit
i,j,k for any index into the dart vector -
get_state_meta_data looks like it does not account for some variables
existing on staggered grids.