update the 'fixsystem' script to take an optional list of files
fix the comments at the top of the fixsystem script to be accurate.
the script has required the fortran compiler name for a while now
but it was still marked as optional.

added code to allow a list of files to be updated after the
compiler name.

simplified the logic to pull the settings for each compiler
out of the loop.  (doesn't change during the loop execution.)

i believe previous versions of this script had a bug in the
nag fortran settings.  this is fixed, but HAS NOT BEEN TESTED.
i've requested access to a machine with NAG to verify this.

this version of the script will not print any output if the
target files are already in the right format.