use the fixsystem in the utilities directory
made one more change to fixsystem to not require the marker strings
to both be present in the source file.  update the Makefile to work
with the common fixsystem.  update the comments at the top of the
source files to say that fixsystem should automatically take care
of the system() interface block.