allow different array types to be passed to MPI calls
NAG seems to enforce that if you make an MPI call with one
type of array and then a different type (e.g. real and then
integer), it is flagged as an error for an inconsistent argument
type, even though MPI has overloaded interfaces for any array type.
the -mismatch flag prints an annoying warning but doesn't stop
your compile.