change several subroutines to pass arguments rather than reference global variables
while rereading the source code for this new test, it was confusing to me
(and i wrote it!) that the check routine was the same for 2 different tests
and took no arguments.  it turns out it was referencing 3 global variables
to change what it did.  i think it is less confusing to pass in arguments explicitly.
i also tried to update some of the comments to be more clear.

i do note that there is still one global array that is referenced in many
places without being explicitly passed but i chose to leave that as-is.