add examples of getting a year's worth of obs
note that the cosblock program is needed.  this is some kind
of standard system utility.  we don't have the source for it.
i made a local copy in the bufr directory in cheyenne only
because the location of it moved around and it was hard to
find it each time.