Initial hybrid scheme implementation
A simple code that does hybrid EnKF-OI DA has been added. The
changes makes it possible to read in a set of static (climatology;
time-invariant) states on top of the regular ensemble restarts in order
to perform hybrid DA.

The climatology states are read in and the forwrad operators
are computed for those states. In filter_mod, there are 2 ensemble
handles; one for the time-evolving ensemble and the other is the
climatology. The same goes for obs_ens handles. In filter_assim,
hybridized obs increments are computed by weighting the ensemble and
clim increments with a factor between 0 and 1. Those hybridized increments
are then used to computed hybrid regression coefficients before
updating the state.

The hybrid weighting factor is currently hard coded and is not adaptive.
Follow-up changes will generalize this procedure such that the weight is read in
from the input.nml and then it gets adaptively updated using the
observations during cycling.