Surface Chlorophyll Obs_Converter
Ocean Color observation converter for NASA's AQUA/MODIS and VIIRS
satellites is now supported. The main program of the converter
can be found in: obs_converters/ocean_color/convert_sat_chl.f90
The script that downloads the data and prepares the files for
conversion can be run from the shell_scripts directory. An
example netcdf data file is available in the data directory.