Support for log transformations
The obs_converter is modified such that the data is always log transformed.
The interface program trans_mitdart.f90 is updated giving the user
the option to transform the bgc variables during the EnKF update.
The clamping procedure in model_mod has been updated to support
logarithm transformations. Also, fixed a bug in model_mod
where 0 values locations were getting masked. This is not true
because trans_mitdart already took care of that.