do the easy fixes first to make model_mod_check work better with MPI
some routines need to be called on all tasks (e.g. read_state, write_state)
but some computations are simply doing the same thing on all tasks
which doesn't provide any new info and makes debugging harder.
and some parts of the code are opening files and writing text to them
which makes for a mess if running with more than one task.

restrict the replicated parts of code to only run on task 0.
fix places where the code opens a text file and writes out lines
so that only happens from a single task.

add a new variable that stores the number of items from the state
vector that are on each task, and use that for looping and allocations.

if running with more than one task comment out test 7 for now
while working on a better fix.

this should work with 1 task or multiple tasks without crashing
or generating unintelligible output.