make the 'closest gridpoint' test work when model_mod_check is built with MPI
make the closest state vector location (gridpoint) test check only the data on
its own task.  if running with MPI, each task then sends the closest distance
and index to task 0, which selects the minimum distance and prints the results.

note that this requires a change in behavior - before if there were multiple
locations with the same distance they'd all be printed out.  now it only prints
the first location it finds that has the minimum distance.

for test 6 this now correctly only prints get_state_meta_data() for the
index values on task 0.  it also writes them to a file.  if metadata for
all items in the state vector are really wanted, the mpi version of the
code will need to change to write a file per task.

this also does not print the entire 'get_state_meta_data()' results to the
screen for test 6.  for large state vectors this is incredibly slow.
the results go to a file as before.

this code required sending single scalars with send_to and receive_from
and also i8s in addition to type real. an overloaded interface was added
to mpi and null_mpi utilities to accomodate this.