remove large chunks of replicated code in the location-specific files
i updated the common code to support MPI but there was replicated code in
one of the location dependent files that i didn't want to update also
since 99% of it didn't have to be different.

isolate the actual differences for different location mods into much smaller
chunks so more common code can be used for all location types.  i have tackled
the single interpolation and find closest gridpoint.  next i will look at the
multiple interpolation and the code that writes a netcdf file but that is
not yet in this commit.

this code which will be easier to maintain if there is only a single place
to update it.

also remove unused use statement routines which were just copied along.