simplified the location dependent code as much as seems reasonable
the test that creates a grid of points and calls interpolate to
make a netcdf file and matlab file for plotting has too many
location dependent differences.  leave in these files for now.

did use our netcdf utility routines to simplify reading the lines
that create the netcdf file output.