Rescale of Climatology prior to assimilation
Added a functionality to rescale the static background covariance
matrix prior to assimilation. The rescaling is useful in the
sense that the rescaled static background covariance can more
appropriately match the total forecast-error variance. The
technique follows Wang et al. (2006); MWR. The new code was
tested both serially and in parallel and the results are
equivalent.