Fixing rank histogram calculations
To calculate the prior and the posterior rank histograms, the
ensemble rank of each of the 40 variables is computed given the
truth. Previously, there was a bug in the calculation that only
displayed the rank histograms for the first 20 variables (when
ens_size=20). The bug fix produces accurate frequency values on
y-axis of the rank histogram plots, however, the overall interpretation
of the results is essentially the same.