move the test for obs "too high" from the interp code to vert convert
as long as all obs are vert converted first, this code change marks
obs which are too high with a QC of 8 and they won't be assimilated.
more importantly, the change allows radiance obs, which must use
interpolate() to build columns of values, to be assimilated.
without this fix all radiance obs will return a QC4 and fail
to be assimilated.