Feature enhancements: Added additional numerical robustness to the normal tails in the RH distribution. This required modified code in both tails for both the transform to and from probit space. The enhancements add additional cost but should prevent possible instances of generating posterior members that violate the bounds.
Also added code to correct input bound violations for
the RH prior. These can be generated in extremely rare cases by roundoff from
the increment computation in the obs space RHF in assim_tools_mod (see
comment there). This is turned on or off by a logical parameter in the code
that is set to off by default. In the short-term, there is a comment in the
error checking that alerts users to the possibility of using this option to
fix the roundoff issue.