A number of independent changes 1. Added namelist control for two parameters, whether to do a logit or probit transform and whether to fix cases where the incoming ensemble members exceed the bounds. 2. Added code to do to_probit and from_probit test every time to_probit_bounded_normal_rh is called. Want these to be very nearly inverses. 3. Corrected imprecise computations of quantiles in to_probit_bounded_normal_rh for the case of a previously computed distribution when in normal tails. 4. Corrected a bug in ens_quantiles that assigned improper quantiles when ensemble members were exactly at the upper bound.
Unfortunately, active debugging is continuing in this commit so there are a bunch of
nasty writes.

Bottom line: This code works while maintaining pretty small inversion differences for the
tracer test case with a positive tracer when assimilating both kinds of obs. It does not
maintain small differences for the case with a negative tracer or for a case with tracer
bounded above at 1. This needs to be fixed.